refactor: update imports to use types from types/editor
