+++ src/api/send-msg.ts
@@ -15,7 +15,8 @@ import { generateMailRequest } from 'store/editor-slice-utils';
 import { getConvEmailStoreAction } from 'store/emails/actions/get-conv-action';
 import { getMessageEmailStoreAction } from 'store/emails/actions/get-message';
 import { getMessageWithExistingParticipantsEmailStoreAction } from 'store/emails/actions/get-message-with-existing-participants';
-import { MailMessage, MailsEditorV2, SaveDraftRequest, SaveDraftResponse } from 'types/index.d';
+import { MailsEditorV2 } from 'types/editor';
+import { MailMessage, SaveDraftRequest, SaveDraftResponse } from 'types/index.d';
 
 export const sendMsg = async ({
 	msg
