+++ src/store/editor/hooks/editor.ts
@@ -8,7 +8,7 @@ import { useCallback, useMemo } from 'react';
 import { useSaveDraftFromEditor } from 'store/editor/hooks/save-draft';
 import { computeAndUpdateEditorStatus, useEditorSetDirty } from 'store/editor/hooks/statuses';
 import { useEditorsStore } from 'store/editor/store';
-import { MailsEditorV2 } from 'types/index.d';
+import { MailsEditorV2 } from 'types/editor';
 
 /**
  * TODO for future refactors
