+++ src/store/editor/hooks/send.ts
@@ -10,10 +10,11 @@ import { find } from 'lodash';
 
 import { sendMsgFromEditor } from 'api/send-msg';
 import { createCancelableTimer } from 'helpers/timers';
-import { computeAndUpdateEditorStatus } from 'store/editor/hooks/statuses';
 import { getEditor } from 'store/editor/hooks/editors';
+import { computeAndUpdateEditorStatus } from 'store/editor/hooks/statuses';
 import { useEditorsStore } from 'store/editor/store';
-import { MailsEditorV2, SaveDraftResponse } from 'types/index.d';
+import { MailsEditorV2 } from 'types/editor';
+import { SaveDraftResponse } from 'types/index.d';
 
 export type SendMessageOptions = {
 	cancelable?: boolean;
