+++ src/store/emails/actions/save-draft-action.ts
@@ -9,7 +9,8 @@ import { API_REQUEST_STATUS } from 'constants/index';
 import { normalizeMailMessageFromSoap } from 'normalizations/normalize-message';
 import { createSoapDraftRequestFromEditor } from 'store/editor/editor-transformations';
 import { updateMessages, updateMessageStatus } from 'store/emails/store';
-import { MailAttachment, MailsEditorV2 } from 'types/index.d';
+import { MailsEditorV2 } from 'types/editor';
+import { MailAttachment } from 'types/soap/save-draft';
 
 type SaveDraftEmailStoreAction = {
 	editor: MailsEditorV2;
