+++ src/views/app/detail-panel/edit/attachment-preview.tsx
@@ -32,7 +32,8 @@ import {
 import { useEditorUploadProcess } from 'store/editor/hooks/updload';
 import { getEditor, useEditorAttachments, useEditorSubject } from 'store/editor/index';
 import StyledWrapper from 'styled-wrapper';
-import { MailsEditorV2, SavedAttachment, UnsavedAttachment } from 'types/index.d';
+import { SavedAttachment, UnsavedAttachment } from 'types/attachments';
+import { MailsEditorV2 } from 'types/editor';
 import { AttachmentUploadStatus } from 'views/app/detail-panel/edit/attachment-upload-status';
 
 const AttachmentHoverBarContainer = styled(Container)`
