+++ src/views/app/detail-panel/edit/parts/edit-view-draft-save-info.tsx
@@ -10,7 +10,7 @@ import { t } from '@zextras/carbonio-shell-ui';
 import moment from 'moment';
 
 import { TIMEOUTS } from 'constants/index';
-import { DraftSaveProcessStatus } from 'types/index.d';
+import { DraftSaveProcessStatus } from 'types/editor';
 import * as StyledComp from 'views/app/detail-panel/edit/parts/edit-view-styled-components';
 
 export const EditViewDraftSaveInfo: FC<{ processStatus?: DraftSaveProcessStatus }> = ({
