+++ src/views/app/detail-panel/edit/parts/tests/text-editor-container.test.tsx
@@ -12,7 +12,7 @@ import { setupTest } from '@test-setup';
 import { useUserSettings } from '@test-utils/carbonio-shell-ui/carbonio-shell-ui';
 import { setupEditorStore } from '__test__/generators/editor-store';
 import { generateNewMessageEditor } from 'store/editor/editor-generators';
-import { MailsEditorV2 } from 'types/index.d';
+import { MailsEditorV2 } from 'types/editor';
 import {
 	TextEditorContainer,
 	TextEditorContainerProps
